backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai>=0.4.0
langchain-google-genai>=1.0.1
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
PyPDF2==3.0.1
pymupdf==1.23.8
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Google Calendar API
google-api-python-client==2.108.0
google-auth==2.23.4

# Hume AI for emotion detection
hume>=0.3.0

# Computer vision for webcam capture
opencv-python==4.8.1.78
numpy==1.24.3

# Additional dependencies for Hume AI
aiohttp>=3.9.0
pillow>=10.0.0
requests>=2.31.0
